Update README.md - reducing section: Creator, servants, prayer and religion
## about the Creator, His servants, prayer and religion
sevenfalling is not a religious organization, nor do We endorse any particular religion or set of beliefs.  when We ask you to #praytoverify it is only to combat mis-information campaigns underway by the c.i.a. and u.s. federal government.  [take a few moments to learn more about the Creator, His servants, prayer and religion](https://github.com/sevenfalling/mission23/wiki/about-the-Creator,-His-servants,-prayer-and-religion).